requirements.txt
pdfplumber==0.11.4
qdrant-client==1.17.1
python-dotenv==1.0.1
pandas==2.2.2
pytesseract==0.3.13
Pillow==10.4.0
tqdm==4.66.4
anthropic==0.96.0
pydantic>=2.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
langchain>=0.3.0
torch>=2.0.0
ddgs>=0.1.0
